Completed_Ability_Feature

+++ src/auto-battle/Assets/Code/Gameplay/Features/Statuses/Configs/StatusSetup.cs
@@ -6,7 +6,7 @@ namespace Code.Gameplay.Features.Statuses.Configs
     public class StatusSetup
     {
         public StatusTypeId TypeId;
-        public int Value;
+        public float Value;
         public int Lifetime;
     }
 }
